# Reimplement `php_round_helper()` using `modf()`