@SakiTakamachi 2023-09-16
@TimWolla 
lgtm!

Since `math.h` may have an issue with the FPU control word, I built your branch on alpine 32bit and ran all the tests.
All tests performed as expected, and we found that `modf` and `fmod` can be used with confidence.