@bukka 2023-09-17
Are you able to do some perf tests to see if there's any difference?