@TimWolla 2023-09-17
With main being:

```
int
main() {
	s[0] = 0xbe0abf86eeacfd2d;
	s[1] = 0x5212a180ba6c1136;
	s[2] = 0xbb1f87b46572ab77;
	s[3] = 0xab0fde1b8ab187da;

	const double step_size = 2.0 / (1ULL << 53);

	double sum = 0;
	for (size_t i = 0; i < 100000000; i++) {
		sum += FUNC(step_size * (next() >> 11) - 1.0, PHP_ROUND_HALF_EVEN);
	}

	printf("%.17g\n", sum);
}
```
(i.e. `(1.0, 1.0)` with rounding half to even “banker's rounding”) my version beats both the old version and Jorg's version:

```
$ gcc test.c -O2 -lm -DFUNC=old -o old
$ gcc test.c -O2 -lm -DFUNC=new_tim -o new_tim
$ gcc test.c -O2 -lm -DFUNC=new_jorg -o new_jorg
$ perf stat ./old
-2015

 Performance counter stats for './old':

          1,667.57 msec task-clock                #    0.999 CPUs utilized          
                37      context-switches          #    0.022 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
                80      page-faults               #    0.048 K/sec                  
     4,855,980,455      cycles                    #    2.912 GHz                      (83.22%)
     1,865,601,989      stalled-cycles-frontend   #   38.42% frontend cycles idle     (83.23%)
       567,348,548      stalled-cycles-backend    #   11.68% backend cycles idle      (66.75%)
     7,204,940,510      instructions              #    1.48  insn per cycle         
                                                  #    0.26  stalled cycles per insn  (83.47%)
     1,001,937,636      branches                  #  600.839 M/sec                    (83.46%)
        50,116,987      branch-misses             #    5.00% of all branches          (83.35%)

       1.669524323 seconds time elapsed

       1.668623000 seconds user
       0.000000000 seconds sys


$ perf stat ./new_jorg
-2015

 Performance counter stats for './new_jorg':

          1,738.67 msec task-clock                #    0.988 CPUs utilized          
               303      context-switches          #    0.174 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
                82      page-faults               #    0.047 K/sec                  
     5,084,530,195      cycles                    #    2.924 GHz                      (83.16%)
     1,972,234,079      stalled-cycles-frontend   #   38.79% frontend cycles idle     (83.38%)
       532,719,936      stalled-cycles-backend    #   10.48% backend cycles idle      (66.80%)
     7,699,914,640      instructions              #    1.51  insn per cycle         
                                                  #    0.26  stalled cycles per insn  (83.63%)
     1,151,052,259      branches                  #  662.029 M/sec                    (83.43%)
        50,208,400      branch-misses             #    4.36% of all branches          (83.24%)

       1.760317642 seconds time elapsed

       1.726577000 seconds user
       0.016023000 seconds sys


$ perf stat ./new_tim
-2015

 Performance counter stats for './new_tim':

          1,602.60 msec task-clock                #    0.999 CPUs utilized          
                10      context-switches          #    0.006 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
                81      page-faults               #    0.051 K/sec                  
     4,675,258,290      cycles                    #    2.917 GHz                      (83.28%)
     1,358,777,298      stalled-cycles-frontend   #   29.06% frontend cycles idle     (83.28%)
       773,307,762      stalled-cycles-backend    #   16.54% backend cycles idle      (66.60%)
     6,712,431,080      instructions              #    1.44  insn per cycle         
                                                  #    0.20  stalled cycles per insn  (83.32%)
     1,249,192,424      branches                  #  779.477 M/sec                    (83.45%)
        53,331,830      branch-misses             #    4.27% of all branches          (83.39%)

       1.604168764 seconds time elapsed

       1.603302000 seconds user
       0.000000000 seconds sys

```

```
$ clang test.c -O2 -lm -DFUNC=old -o old
$ clang test.c -O2 -lm -DFUNC=new_jorg -o new_jorg
$ clang test.c -O2 -lm -DFUNC=new_tim -o new_tim
$ perf stat ./old
-2015

 Performance counter stats for './old':

          1,380.73 msec task-clock                #    0.999 CPUs utilized          
                 5      context-switches          #    0.004 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
                84      page-faults               #    0.061 K/sec                  
     3,995,933,536      cycles                    #    2.894 GHz                      (83.20%)
     1,394,148,218      stalled-cycles-frontend   #   34.89% frontend cycles idle     (83.20%)
       862,300,694      stalled-cycles-backend    #   21.58% backend cycles idle      (66.81%)
     4,489,172,806      instructions              #    1.12  insn per cycle         
                                                  #    0.31  stalled cycles per insn  (83.49%)
       900,959,583      branches                  #  652.525 M/sec                    (83.49%)
        53,362,571      branch-misses             #    5.92% of all branches          (83.30%)

       1.381509128 seconds time elapsed

       1.381467000 seconds user
       0.000000000 seconds sys


$ perf stat ./new_jorg
-2015

 Performance counter stats for './new_jorg':

          1,592.40 msec task-clock                #    0.995 CPUs utilized          
                50      context-switches          #    0.031 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
                82      page-faults               #    0.051 K/sec                  
     4,622,058,007      cycles                    #    2.903 GHz                      (83.28%)
     1,458,687,467      stalled-cycles-frontend   #   31.56% frontend cycles idle     (83.18%)
       851,065,229      stalled-cycles-backend    #   18.41% backend cycles idle      (66.83%)
     5,955,613,420      instructions              #    1.29  insn per cycle         
                                                  #    0.24  stalled cycles per insn  (83.45%)
     1,053,075,802      branches                  #  661.313 M/sec                    (83.44%)
        56,444,889      branch-misses             #    5.36% of all branches          (83.27%)

       1.600838210 seconds time elapsed

       1.593224000 seconds user
       0.000000000 seconds sys


$ perf stat ./new_tim
-2015

 Performance counter stats for './new_tim':

          1,298.35 msec task-clock                #    0.997 CPUs utilized          
                53      context-switches          #    0.041 K/sec                  
                 1      cpu-migrations            #    0.001 K/sec                  
                83      page-faults               #    0.064 K/sec                  
     3,773,061,211      cycles                    #    2.906 GHz                      (83.21%)
       989,454,803      stalled-cycles-frontend   #   26.22% frontend cycles idle     (83.39%)
       784,657,116      stalled-cycles-backend    #   20.80% backend cycles idle      (66.75%)
     5,598,290,346      instructions              #    1.48  insn per cycle         
                                                  #    0.18  stalled cycles per insn  (83.37%)
       901,733,975      branches                  #  694.524 M/sec                    (83.43%)
        54,266,355      branch-misses             #    6.02% of all branches          (83.21%)

       1.302192036 seconds time elapsed

       1.299178000 seconds user
       0.000000000 seconds sys
```